refactor: move createFactory helper to new file
