+++ src/index.js
@@ -1,4 +1,5 @@
-import { factory, defaultConfig } from './block';
+import factory from './createFactory';
+import { defaultConfig } from './block';
 import { getDataProps } from './data';
 
 export { factory, getDataProps, defaultConfig };
